pyproject.toml
[project]
name = "learn-python-grading-journal"
version = "0.1.0"
description = "School digital grading book service backend created for learning purposes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.115.6",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
]
